Handle 'empty IMSI' records in web backend
Summary:
In the course of local (Vagrant) testing I discovered that funky
subscriber records were being created and sent to the cloud. These had no
IMSI associated, which caused the web UI to break when trying to lookup
the IMSI associated with the subscriber. There's a deeper question of what
exactly happened there, but this change suffices it to prevent the UI being
broken in that case.

Reviewed By: shaddi

Differential Revision: D4177571

fbshipit-source-id: 4220900ba126faf6e1329e12b655c92c3cd684af