Add option to set number kind
Summary:
When adding numbers to the db it's necessary to set the kind field so
that Freeswitch handles call routing correctly. The default is to add numbers as
'local' rather than 'number.telecom.permanent' so that we don't accidentally send
a load of junk traffic to the telecom partner's SIP gateway.

Reviewed By: kkroo

Differential Revision: D4627762

fbshipit-source-id: 601d5224a4c6493f9117cb5866292e4f5c795242