Make Stripe credentials optional
Summary:
It shouldn't be necessary to enter Stripe credentials (even fake
ones) in order to create a new network or change an existing one. If I had
any JS Fu I'd have figured out how to disable/hide the associated fields,
which are only required when auto-reload is enabled, but I didn't.

Reviewed By: shaddi

Differential Revision: D4645116

fbshipit-source-id: 9660d55a0985b730cd46ede302512c5fe1910ca4