Added git support to fabric
Summary:
Allow fabric to fall back to git if a mercurial repo is not found or mercurial is not installed. This is needed for a real deployment of the git based public projected. As with the previous pull request, please let us know what the standards are for community pull requests if there is anything we should update or change!
Cheers,
-mj
Closes https://github.com/facebookincubator/CommunityCellularManager/pull/26

Reviewed By: kkroo

Differential Revision: D4463040

Pulled By: 9muir

fbshipit-source-id: 769dd2a74112ea9d9b9f97947c5cbb2bdc1fc570