Backed out changeset 8871e7837fc9
Summary:
The summary of diff D5072969 is not representative of the actual code changes.
At first this didn't concern me as much as it should have because the evolution
of the diff was captured in the comments during the review process.

However, after the diff landed, it was also pushed to an external open source repo
which only captures that first summary, which is less inaccurate.

Reviewed By: kkroo

Differential Revision: D5410793

fbshipit-source-id: f70fe396444f87650ed340c7f773b7eadd7d9e75