export hangup cause and put it in CDR reason/description
Summary:
This gets the [hangup_cause value](https://wiki.freeswitch.org/wiki/Hangup_Causes) from the FS xml cdrs and includes them in the cdr description.
Our intention for this is to get more info from calls, especially those zero-second calls that we see in the CDRs. This might be useful for others as well.

CDR reason/description will now be in this format:
```
2 sec call to 639361234567 (outside_call/NORMAL_CLEARING)
0 sec call to 639361234567 (local_call/USER_BUSY)
0 sec call to 639361234567 (outside_call/RECOVERY_ON_TIMER_EXPIRE)
```
Closes https://github.com/facebookincubator/CommunityCellularManager/pull/93

Reviewed By: 9muir

Differential Revision: D7110623

Pulled By: shaddi

fbshipit-source-id: 7c0cc92b72a2bb459806e208a78958447bd16cd1